Make tileflag names consistent and intuitive
- `TILEFLAG_VFLIP` -> TILEFLAG_FLIP_HORIZONTAL
- `TILEFLAG_HFLIP` -> TILEFLAG_FLIP_VERTICAL

According to the native editor, the "Tiled" editor and image search, a
horizontal flip should be associated with switching left and right, modifying
the x coordinate.

I did not just switch the letters `H` and `V` to create compiler errors
where the original constants are used.

Whenever I was working with tileflags, the naming caused me to have no
idea what I was doing. I mostly had to resort to opening the resulting
map in the editor to see what the code does. This change aims to make
the naming intuitive and also consistent with the map editor.