mastersrv: Implement QUIC challenge
On register, the server sends its public key, as well as a private key
with which the master server should initiate the QUIC connection for the
challenge.
The masterserver establishes the QUIC connection with a certificate
derived from the given private key and verifies the certificate of the
connection with the given public key of the server.